requirements.txt
adrf==0.1.2
asgiref==3.7.2
async-property==0.2.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.9.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.2.0
click==8.1.7
cloudinary==1.33.0
colorama==0.4.6
constantly==23.10.4
cryptography==41.0.5
daphne==4.0.0
defusedxml==0.7.1
Django==4.2.3
django-autoslug==1.9.9
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-countries==7.5.1
django-debug-toolbar==4.1.0
django-filter==23.4
django-jazzmin==2.6.0
django-phonenumber-field==7.2.0
django-redis==5.4.0
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf-yasg==1.21.7
exceptiongroup==1.2.0
google-ai-generativelanguage==0.4.0
google-api-core==2.15.0
google-auth==2.25.2
google-generativeai==0.3.1
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
h11==0.14.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
msgpack==1.0.7
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.2
phonenumberslite==8.13.26
Pillow==10.1.0
platformdirs==3.10.0
pluggy==1.3.0
proto-plus==1.23.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==1.7.1
pymemcache==4.0.0
pyOpenSSL==23.3.0
pytest==7.4.3
python-decouple==3.8
python-memcached==1.59
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.13.1
rsa==4.9
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.4
stripe==7.7.0
tomli==2.0.1
tqdm==4.66.1
treblle==0.14
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.25.0
whitenoise==6.6.0
zope.interface==6.1
